+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -1413,6 +1413,7 @@ class FrameworkExtension extends Extension
             $container->removeDefinition('console.command.translation_extract');
             $container->removeDefinition('console.command.translation_pull');
             $container->removeDefinition('console.command.translation_push');
+            $container->removeDefinition('console.command.translation_xliff_update_sources');
 
             return;
         }
@@ -1488,6 +1489,12 @@ class FrameworkExtension extends Extension
 
         if ($container->hasDefinition('console.command.translation_extract')) {
             $container->getDefinition('console.command.translation_extract')->replaceArgument(6, $transPaths);
+        }
+
+        if ($container->hasDefinition('console.command.translation_xliff_update_sources')) {
+            $container->getDefinition('console.command.translation_xliff_update_sources')
+                ->replaceArgument(3, array_merge($transPaths, [$config['default_path']]))
+            ;
         }
 
         if (null === $defaultDir) {