+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/console.php
@@ -57,6 +57,7 @@ use Symfony\Component\Serializer\Command\DebugCommand as SerializerDebugCommand;
 use Symfony\Component\Translation\Command\TranslationPullCommand;
 use Symfony\Component\Translation\Command\TranslationPushCommand;
 use Symfony\Component\Translation\Command\XliffLintCommand;
+use Symfony\Component\Translation\Command\XliffUpdateSourcesCommand;
 use Symfony\Component\Validator\Command\DebugCommand as ValidatorDebugCommand;
 
 return static function (ContainerConfigurator $container) {
@@ -304,6 +305,16 @@ return static function (ContainerConfigurator $container) {
                 [], // Enabled locales
             ])
             ->tag('console.command', ['command' => 'translation:push'])
+
+        ->set('console.command.translation_xliff_update_sources', XliffUpdateSourcesCommand::class)
+            ->args([
+                service('translation.writer'),
+                service('translation.reader'),
+                param('kernel.default_locale'),
+                [], // Translator paths
+                param('kernel.enabled_locales'),
+            ])
+        ->tag('console.command')
 
         ->set('console.command.workflow_dump', WorkflowDumpCommand::class)
             ->args([