+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
  * Add `secrets:reveal` command
  * Add `rate_limiter` option to `http_client.default_options` and `http_client.scoped_clients`
  * Attach the workflow's configuration to the `workflow` tag
+ * Add a `translation:update-xliff-sources` command to use de default locale target in the `source` tag of XLIFF files 
 
 7.0
 ---